requirements.txt
requests
python-dotenv
plotly
pandas
dash